// frontend/src/api/generated/serializers/BillingOverview.ts
// Typelizer digest f595eaca64d087ce5c084d477f437712
//
// DO NOT MODIFY: This file was automatically generated by Typelizer.
import type {Subscription, Offer} from "./index"

type BillingOverview = {
  plan: string;
  offerRevision: string;
  sales: "open" | "test" | "closed";
  canManage: boolean;
  subscription: Subscription | null;
  offers: Array<Offer>;
}

export type { BillingOverview };