// frontend/src/api/generated/serializers/Subscription.ts
// Typelizer digest 46f27a75861776c7374508561cf49b89
//
// DO NOT MODIFY: This file was automatically generated by Typelizer.

type Subscription = {
  plan: string;
  offerId: string | null;
  status: string;
  cancelAtPeriodEnd: boolean;
  paused: boolean;
  currentPeriodEnd: string | null;
  paid: boolean;
}

export type { Subscription };